requirements.txt
scrapy
elasticsearch
scrapy_proxies
python-dateutil
python-arango